Rename SourcesPage to just Sources (it's already in pages)
