+++ src/Routes.js
@@ -2,7 +2,7 @@ import React, { Suspense, lazy } from 'react';
 import { Route } from 'react-router-dom';
 import ContentLoader from 'react-content-loader';
 
-const SourcesPage = lazy(() => import(/* webpackChunkName: "sourcePage" */ './pages/SourcesPage'));
+const SourcesPage = lazy(() => import(/* webpackChunkName: "sourcePage" */ './pages/Sources'));
 
 const Loader = () => (
     <ContentLoader>
